[ANDROAPP-4621] Remove EventDetailsRepository as interface
